requirements.txt
absl-py==1.4.0
 asttokens==2.4.0
 backcall==0.2.0
 cachetools==5.3.1
 certifi==2023.7.22
 charset-normalizer==3.2.0
 contourpy==1.1.0
 cycler==0.11.0
 decorator==5.1.1
 easing-functions==1.0.4
 einops==0.7.0
 exceptiongroup==1.1.3
 executing==1.2.0
 faiss-cpu==1.7.3
 fonttools==4.42.1
 google-auth==2.22.0
 google-auth-oauthlib==1.0.0
 grpcio==1.57.0
 idna==3.4
 imageio==2.31.3
 importlib-metadata==6.8.0
 importlib-resources==6.0.1
ipython==8.15.0
 jedi==0.19.0
 kiwisolver==1.4.5
 lazy_loader==0.3
 Markdown==3.4.4
 MarkupSafe==2.1.3
 matplotlib==3.7.2
 matplotlib-inline==0.1.6
 networkx==3.1
 numpy==1.25.2
 oauthlib==3.2.2
 opencv-python==4.8.0.74
 packaging==23.1
 pandas==2.1.0
 parso==0.8.3
 pexpect==4.8.0
 pickleshare==0.7.5
 Pillow==10.0.0
 prompt-toolkit==3.0.39
 protobuf==4.24.2
 ptyprocess==0.7.0
 pure-eval==0.2.2
 pyasn1==0.5.0
 pyasn1-modules==0.3.0
 Pygments==2.16.1
 pyparsing==3.0.9
 python-dateutil==2.8.2
 pytz==2023.3.post1
 PyWavelets==1.4.1
 requests==2.31.0
 requests-oauthlib==1.3.1
 rsa==4.9
 scikit-image==0.21.0
 scipy==1.11.1
 six==1.16.0
 stack-data==0.6.2
 tensorboard==2.13.0
 tensorboard-data-server==0.7.1
 tifffile==2023.8.30
 torch==1.13.1+cu116
 torchinfo==1.8.0
 torchvision==0.14.1+cu116
 tqdm==4.66.1
 traitlets==5.10.0
 typing_extensions==4.7.1
 tzdata==2023.3
urllib3==1.26.16
 wcwidth==0.2.6
 Werkzeug==2.3.7
 zipp==3.16.2